pyproject.toml
[tool.poetry]
name = "australia-microplastic"
version = "0.1.0"
description = ""
authors = ["Alan Huynh <hdmquan@outlook.com>"]
readme = "README.md"
packages = [{include = "src", from = "."}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.6.0"
holoviews = "^1.20.1"
pandas = "^2.2.3"
numpy = "^2.2.3"
loguru = "^0.7.3"
bokeh = "^3.6.3"
scipy = "^1.15.2"
fastparquet = "^2024.11.0"
plotly = "^6.0.0"
nbformat = "^5.10.4"
statsmodels = "^0.14.4"
dotenv = "^0.9.9"
modis-tools = "^1.2.0"
python-cmr = "^0.13.0"
rasterio = "^1.4.3"
pyhdf = "^0.11.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
